split spec extensions into extension classes
