+++ src/blueprints/Do.Blueprints.Service.Application/Extensions/ServiceSpecDictionaryExtensions.cs
@@ -1,0 +1,11 @@
+using Do.Testing;
+
+namespace Do;
+
+public static class ServiceSpecDictionaryExtensions
+{
+    public static Dictionary<string, string> ADictionary(this Stubber giveMe) => giveMe.ADictionary<string, string>();
+    public static Dictionary<TKey, TValue> ADictionary<TKey, TValue>(this Stubber _, params (TKey, TValue)[] pairs)
+        where TKey : notnull
+    => pairs.ToDictionary(pair => pair.Item1, pair => pair.Item2);
+}
