+++ src/blueprints/Do.Blueprints.Service.Application/Extensions/ServiceSpecGuidExtensions.cs
@@ -1,0 +1,17 @@
+using Do.Testing;
+
+namespace Do;
+
+public static class ServiceSpecGuidExtensions
+{
+    public static Guid AGuid(this Stubber _,
+        string? starts = default
+    )
+    {
+        starts ??= string.Empty;
+
+        const string template = "4d13bbe0-07a4-4b64-9d31-8fef958fbef1";
+
+        return Guid.Parse($"{starts}{template[starts.Length..]}");
+    }
+}
