+++ src/blueprints/Do.Blueprints.Service.Application/Extensions/ServiceSpecIntegerExtensions.cs
@@ -1,0 +1,8 @@
+using Do.Testing;
+
+namespace Do;
+
+public static class ServiceSpecIntegerExtensions
+{
+    public static int AnInteger(this Stubber _) => 42;
+}
